+++ src/main/java/miyucomics/hexical/casting/patterns/eval/OpJanus.kt
@@ -7,5 +7,5 @@ import at.petrak.hexcasting.api.spell.casting.eval.SpellContinuation
 import at.petrak.hexcasting.api.spell.iota.Iota
 
 object OpJanus : Action {
-	override fun operate(continuation: SpellContinuation, stack: MutableList<Iota>, ravenmind: Iota?, ctx: CastingContext) = OperationResult(SpellContinuation.Done, listOf(), null, listOf())
+	override fun operate(continuation: SpellContinuation, stack: MutableList<Iota>, ravenmind: Iota?, ctx: CastingContext) = OperationResult(SpellContinuation.Done, listOf(), ravenmind, listOf())
 }