+++ src/main/java/miyucomics/hexical/interfaces/FrameForEachMinterface.kt
@@ -4,4 +4,5 @@ import miyucomics.hexical.enums.InjectedGambit
 
 interface FrameForEachMinterface {
 	fun overwrite(gambit: InjectedGambit)
+	fun getInjectedGambit(): InjectedGambit
 }